models.py submit full o' errors
An attempt at creating the models.py. Some or all of it is wrong,
especially the ForeignKey and ManyToManyField choices. Also just for the
exercise I included stuff we don't need in the MVP, like the
Administrator class.

If you ever get a chance, it'd be great if you could correct this with
me over VoIP (maybe for like 15-30 mins). In the meantime, I'm going to
start a Django tutorial, so no rush. Maybe the tutorial will help me
figure out what I did right and wrong in this models.py.